Rip Path and GPath into bolt module
The big one. Turned out to be much less painless than I thought it would
be. The one big glaring question mark is the TODO I left at the top of
the new paths.py.

Many of these import changes will obviously have to be reverted once the
bolt module is done, but that can be done in a separate commit.